+++ Sources/StreamChat/WebSocketClient/Events/Event.swift
@@ -7,7 +7,7 @@ import Foundation
 /// An `Event` object representing an event in the chat system.
 public protocol Event {}
 
-extension Event {
+public extension Event {
     var name: String {
         String(describing: Self.self).replacingOccurrences(of: "DTO", with: "")
     }