+++ Tests/StreamChatTests/WebSocketClient/Events/MessageEvents_Tests.swift
@@ -82,6 +82,34 @@ final class MessageEvents_Tests: XCTestCase {
         XCTAssertEqual(event?.hardDelete, true)
     }
 
+    func test_messageDeletedEvent_toDomainEvent() throws {
+        let json = XCTestCase.mockData(fromJSONFile: "MessageDeleted")
+        let event = try eventDecoder.decode(from: json) as? MessageDeletedEventDTO
+
+        let channelId = try XCTUnwrap(event?.cid)
+        let message = try XCTUnwrap(event?.message)
+        let session = DatabaseContainer_Spy(kind: .inMemory).viewContext
+        _ = try session.saveChannel(payload: .dummy(cid: channelId), query: nil, cache: nil)
+        _ = try session.saveMessage(payload: message, for: channelId, cache: nil)
+
+        let domainEvent = event?.toDomainEvent(session: session)
+        XCTAssertEqual(domainEvent is MessageDeletedEvent, true)
+    }
+
+    func test_messageDeletedEvent_toDomainEvent_whenIsHardDeleted_whenMessageNotInLocalDB() throws {
+        let json = XCTestCase.mockData(fromJSONFile: "MessageDeletedHard")
+        let event = try eventDecoder.decode(from: json) as? MessageDeletedEventDTO
+
+        let channelId = try XCTUnwrap(event?.cid)
+        let message = try XCTUnwrap(event?.message)
+        let session = DatabaseContainer_Spy(kind: .inMemory).viewContext
+        // Only save the channel. Not the message. In this case the payload should be directly mapped to model.
+        _ = try session.saveChannel(payload: .dummy(cid: channelId), query: nil, cache: nil)
+
+        let domainEvent = try XCTUnwrap(event?.toDomainEvent(session: session) as? MessageDeletedEvent)
+        XCTAssertEqual(domainEvent.message.id, message.id)
+    }
+
     func test_read() throws {
         let json = XCTestCase.mockData(fromJSONFile: "MessageRead")
         let event = try eventDecoder.decode(from: json) as? MessageReadEventDTO